Feature/#10 staff>order

+++ src/main/java/com/linasdeli/api/domain/Order.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer oid;
+    private Long oid;
 
     @ManyToOne
     @JoinColumn(name = "platter_id", nullable = false)
